Search: Extract visitor helper functions into utils
